refactor(session): split rollback helper out of trace_event_chunks
Keeps trace_event_chunks.py within the 1000-line backend source limit;
TraceEventChunkMixin now inherits TraceChunkRollbackMixin.
